+++ extension/custom/product/ext/config/config-ext-product.php
@@ -7,6 +7,7 @@ $config->product->search['fields']['purchaser']     = $lang->story->purchaser;
 $config->product->search['fields']['responseResult']     = $lang->story->responseResult;
 
 
+
 $config->product->search['params']['bzCategory']         = array('operator' => '=',       'control' => 'select', 'values' => ( array(''=>'') + $lang->story->bzCategoryList) );
 $config->product->search['params']['prCategory']         = array('operator' => '=',       'control' => 'select', 'values' => (array(''=>'') + $lang->story->prCategoryList) );
 $config->product->search['params']['uatDate']     = array('operator' => '=', 'control' => 'input', 'values' => '', 'class' => 'date');
@@ -20,3 +21,15 @@ $config->product->search['params']['prLevel']     = array('operator' => '=', 'co
 $config->product->search['fields']['bizProject']     = $lang->story->bizProject;
 $config->product->search['params']['bizProject']     = array('operator' => '=', 'control' => 'select', 'values' => 'bizProjects');
 
+$config->product->search['fields']['asort']     = $lang->story->asort;
+$config->product->search['params']['asort']     = array('operator' => 'include', 'control' => 'input');
+
+
+$config->product->search['fields']['rspRecievedTime']     = $lang->story->rspAcceptTime;
+$config->product->search['params']['rspRecievedTime']     = array('operator' => '>=', 'control' => 'input', 'values' => '', 'class' => 'date');
+$config->product->search['fields']['rspRejectTime']     = $lang->story->rspAcceptTime;
+$config->product->search['params']['rspRejectTime']     = array('operator' => '>=', 'control' => 'input', 'values' => '', 'class' => 'date');
+$config->product->search['fields']['rspResearchTime']     = $lang->story->rspAcceptTime;
+$config->product->search['params']['rspResearchTime']     = array('operator' => '>=', 'control' => 'input', 'values' => '', 'class' => 'date');
+$config->product->search['fields']['rspAcceptTime']     = $lang->story->rspAcceptTime;
+$config->product->search['params']['rspAcceptTime']     = array('operator' => '>=', 'control' => 'input', 'values' => '', 'class' => 'date');